# Optimizations to string and assocarr