These give a ~17% speedup when building [sm_disassembly](https://github.com/InsaneFirebat/sm_disassembly) (lightly patched to work with 2.0.0)

```
hyperfine -w 1 --prepare "python tools/ff_file.py ../SM.sfc" --reference "C:\Users\yuriks\projects\asar\build_asar2\asar\bin\RelWithDebInfo\asar.exe --no-title-check --symbols=wla --symbols-path=symbols.sym src/main.asm SM.sfc" "tools\asar2\asar --no-title-check --symbols=wla --symbols-path=symbols.sym src/main.asm SM.sfc"
Benchmark 1: C:\Users\yuriks\projects\asar\build_asar2\asar\bin\RelWithDebInfo\asar.exe --no-title-check --symbols=wla --symbols-path=symbols.sym src/main.asm SM.sfc
  Time (mean ± σ):      2.349 s ±  0.012 s    [User: 1.958 s, System: 0.385 s]
  Range (min … max):    2.327 s …  2.365 s    10 runs

Benchmark 2: tools\asar2\asar --no-title-check --symbols=wla --symbols-path=symbols.sym src/main.asm SM.sfc
  Time (mean ± σ):      2.745 s ±  0.021 s    [User: 2.293 s, System: 0.432 s]
  Range (min … max):    2.720 s …  2.788 s    10 runs

Summary
  C:\Users\yuriks\projects\asar\build_asar2\asar\bin\RelWithDebInfo\asar.exe --no-title-check --symbols=wla --symbols-path=symbols.sym src/main.asm SM.sfc ran
    1.17 ± 0.01 times faster than tools\asar2\asar --no-title-check --symbols=wla --symbols-path=symbols.sym src/main.asm SM.sfc
```