@yuriks 2025-02-18
Seems like the CI failed to build. I didn't get any issues locally with VS 2022, but I'll have to look into it tomorrow.