@yuriks 2025-02-18
The C# test is failing because the order in which labels are returned was different. Is the order of the labels in the API something that is expected to be kept stable, or should I just update the test? Otherwise, I had a previous implementation with `std::map` that I can use instead so they're returned in sorted order.