@randomdude999 2025-02-18
i think it would be nice to return sorted data from the api functions. but my preferred implementation would be to just call sort() on the output before returning it, it'd probably be more efficient than using a sorted map in the rest of asar.

also, it'd probably be interesting to see if using a better hashmap implementation (i think absl's one was good?) would give noticeable speedups.