@randomdude999 2025-02-18
oh and CI is failing on linux because you're missing `#include <cstring>` or something. i guess msvc is more lenient about standard includes