@randomdude999 2025-02-19
oh right, the symbol file also contains all labels, so you'd need to do the sort() [somewhere here](https://github.com/RPGHacker/asar/blob/5e0cdba7566438be20abf4591633c01d83750dd9/src/asar/main.cpp#L926) too.

i hope that that's the last one...