@yuriks 2025-02-19
Yeah, I identified that the symbols output would change, but also didn't see there were separate tests for that. I made it sort by address rather than label name, which I think is generally what one would expect. This should be the last one, all the other uses of `.each` didn't care about ordering.